+++ pymsis/__init__.py
@@ -3,9 +3,9 @@
 import importlib.metadata
 
 from pymsis.msis import Variable, calculate
-from pymsis.utils import set_space_weather_path
+from pymsis.utils import use_space_weather_file
 
 
 __version__ = importlib.metadata.version("pymsis")
 
-__all__ = ["Variable", "__version__", "calculate", "set_space_weather_path"]
+__all__ = ["Variable", "__version__", "calculate", "use_space_weather_file"]