+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
     at higher temporal resolutions. There can be step changes of more than
     20% in some situations when going from hour 02:59 to 03:00 when new
     indices get used again on the 3-hourly boundaries.
-- **ADDED** `pymsis.set_space_weather_path()` function.
+- **ADDED** `pymsis.use_space_weather_file()` function.
   - The function points pymsis to a custom space weather
     file instead of the one bundled in the install location. 
   - The same path can be set at startup with the `PYMSIS_SPACE_WEATHER_PATH`